+++ src/libretro/drivers/camera/__init__.py
@@ -1,2 +1,2 @@
 from .driver import *
-from .generator import *
+from .iterable import *